move robot and sharp
