+++ src/client-electron/worker/worker.ts
@@ -3,11 +3,17 @@ import {
   BreachProtocolDaemonsFragment,
   BreachProtocolGridFragment,
   BreachProtocolOCRFragment,
-  SharpImageContainer,
 } from '@/core';
+import { PlatformRobot, SharpImageContainer } from '@/platform-node';
 import { ipcRenderer as ipc, IpcRendererEvent } from 'electron';
 import { listDisplays, ScreenshotDisplayOutput } from 'screenshot-desktop';
 import sharp from 'sharp';
+import {
+  AddHistoryEntryAction,
+  SetDisplaysAction,
+  SetStatusAction,
+  UpdateSettingsAction,
+} from '../actions';
 import {
   Action,
   AppSettings,
@@ -17,14 +23,7 @@ import {
   workerAsyncRequestListener,
   WorkerStatus,
 } from '../common';
-import {
-  AddHistoryEntryAction,
-  SetDisplaysAction,
-  SetStatusAction,
-  UpdateSettingsAction,
-} from '../actions';
 import { BreachProtocolAutosolver } from './autosolver';
-import { PlatformRobot } from './robot';
 
 export class BreachProtocolWorker {
   private disposeAsyncRequestListener: () => void = null;
@@ -97,9 +96,7 @@ export class BreachProtocolWorker {
 
     this.updateStatus(WorkerStatus.Working);
 
-    const { activeDisplayId } = this.settings;
-    const { dpiScale } = this.displays.find((d) => d.id === activeDisplayId);
-    const robot = new PlatformRobot(this.settings, dpiScale);
+    const robot = this.getRobot();
     const bpa = new BreachProtocolAutosolver(this.settings, robot);
     const entry = await bpa.solve();
 
@@ -114,6 +111,13 @@ export class BreachProtocolWorker {
     if (type === 'UPDATE_SETTINGS') {
       this.settings = payload.settings;
     }
+  }
+
+  private getRobot() {
+    const { activeDisplayId } = this.settings;
+    const { dpiScale } = this.displays.find((d) => d.id === activeDisplayId);
+
+    return new PlatformRobot(this.settings, dpiScale);
   }
 
   private async handleAsyncRequest(req: Request) {
